+++ cmd/config.yaml
@@ -22,6 +22,7 @@ projector:
   kafka:
     bootstrapServers: localhost:9092
     topic: todo
+    consumerGroup: todo
 query:
   entityStoreDB:
     host: localhost
