+++ docker/mongo-init.js
@@ -2,7 +2,7 @@ db.createUser({
     user: "todo_account",
     pwd: "1234",
     roles: [
-        { role: "readWrite", db: "todo_db" }
+        {role: "readWrite", db: "todo_db"}
     ]
 });
