+++ readme.md
@@ -76,7 +76,7 @@ go get github.com/google/uuid
 go get github.com/Shopify/sarama
 ```
 
-###  mongodb
+### mongodb
 
 ```bash
 go get go.mongodb.org/mongo-driver/mongo
